类型迁移；建立网点服务类；阶段性计划

+++ src/main/java/org/example/express_backend/entity/Employee.java
@@ -12,5 +12,5 @@ public class Employee {
     private String email;
     private String passwordHash;
     private String salt;
-    private Integer serveAt;
+    private Long serveAt;
 }
